requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cfgv==3.4.0
colorama==0.4.6
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.6
django-cors-headers==4.3.0
django-crispy-forms==2.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.20.1
filelock==3.12.4
gunicorn==21.2.0
identify==2.5.30
nodeenv==1.8.0
packaging==23.2
Pillow==10.0.1
pipenv==2023.10.3
platformdirs==3.11.0
pre-commit==3.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
virtualenv==20.24.5
virtualenv-clone==0.5.7
whitenoise==6.6.0
